Create Code Sample 1
Explanation of the code: 
This is a JavaScript file that contains functions and components for building forms in a React application. The code imports various libraries, such as React, React Router, and Recoil, as well as custom modules from the project. 

The main function of this code is to map a configuration object that defines the structure of a form into a series of React components that can be rendered in the user interface. The configuration object is passed in as the fieldset parameter to the FormComponent component. 

The mapFieldset function is a recursive function that takes a fieldset object and maps each of its elements to a React component. Each element in the fieldset object can be either a field component or another nested fieldset object, which contains a collection of fields or more nested fieldsets. 

The mapFieldComponent function takes a single field object and maps it to a React component. It does this by looking up the appropriate field mapper function from the fieldMappers object and passing the field object and some additional props to it. 

The translateLocalizableFieldProps function takes a fieldProps object and translates any properties that have been marked for localization using the _ext function from the localization module. 

The asReadOnlyForm function takes a fieldset object and maps it to a read-only version of the form. This is used for displaying the form when it is not editable. 

The FormComponent component is the main component that is rendered when the form is displayed. It takes several props, including the formBuilderState, which contains information about the state of the form, the fieldset object, which defines the structure of the form, and the watcher object, which contains information about the validation state of the form. 

The Intermediate component is a higher-level component that wraps the FormComponent component and provides additional functionality, such as form validation and navigation between pages of the form. It takes several props, including the formBuilderState, which contains information about the state of the form, and the validationConfig, which contains information about the validation rules for the form. It also takes a pages array, which defines the structure of the form, and a readOnly flag, which determines whether the form is editable or not. 

Overall, this code provides a flexible and extensible way to define and render forms in a React application.